test: Fix implicit conversion (UBSan warning)

+++ test/unittests/execute_numeric_test.cpp
@@ -1014,7 +1014,7 @@ TEST(execute_numeric, i64_div_u_by_zero)
 
 TEST(execute_numeric, i64_rem_s)
 {
-    EXPECT_RESULT(execute_binary_operation(Instr::i64_rem_s, uint64_t(-4242), 4200), -42);
+    EXPECT_RESULT(execute_binary_operation(Instr::i64_rem_s, uint64_t(-4242), 4200), uint64_t(-42));
     constexpr auto i64_min = std::numeric_limits<int64_t>::min();
     EXPECT_RESULT(execute_binary_operation(Instr::i64_rem_s, uint64_t(i64_min), uint64_t(-1)), 0);
 }
